# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Nano backend developed in golang.",
        "title": "Swagger nano backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
        "/getBalance/{address}": {
            "get": {
                "description": "return a list of locations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "fetches all locations by ip",
                "operationId": "/getBalance/:address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Address of the wallet",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "Invalid media type, must be application/json",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getGeoLocations": {
            "get": {
                "description": "return a list of locations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "fetches all locations by ip",
                "operationId": "/getGeoLocations",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "Invalid media type, must be application/json",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getPeers": {
            "get": {
                "description": "return a list of peers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "fetches all peers",
                "operationId": "/getPeers",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Page not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "415": {
                        "description": "Invalid media type, must be application/json",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}